Support for immutable ConfigMaps (fixes #13) (#14)
ConfigMaps created to contain mcrouter configuration are now created as immutable. Each time endpoints for memcached clusters change a new configmap is created and the mcrouter deployment is updated. This triggers a rolling update to a new replicaset. The configmap is then assigned the replicaset as it's owner. Old configmaps are garbage collected when the replicaset that references them is garbage collected by the deployment controller.